pyproject.toml
# SPDX-License-Identifier: GPL-3.0-or-later
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-qes-signer"
version = "0.3.0"
description = "GUI tool for visually placing signature fields in PDFs and applying qualified electronic signatures (QES) via PKCS#11 / smartcard, specifically developed for Telesec TCOS 3.0 SigG cards"
readme = "README.md"
license = { text = "GPL-3.0-or-later" }
authors = [
    { name = "Peter Muß", email = "codeberg.org@muss.eu" },
]
keywords = ["pdf", "signature", "qes", "pkcs11", "smartcard", "tcos", "telesec"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Security :: Cryptography",
]
requires-python = ">=3.9"
dependencies = [
    "pymupdf>=1.23",
    "Pillow>=10.0",
    "PyQt6>=6.4",
    "pyhanko>=0.20",
    "pyhanko-certvalidator>=0.26",
    "certifi",
    "python-pkcs11>=0.7",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine",
    "pdoc",
]

[project.scripts]
pdf-signer = "pdf_signer.main:main"

[project.urls]
Homepage = "https://codeberg.org/pitbo/pdf-qes-signer"
Repository = "https://codeberg.org/pitbo/pdf-qes-signer.git"
"Bug Tracker" = "https://codeberg.org/pitbo/pdf-qes-signer/issues"

[tool.setuptools.packages.find]
include = ["pdf_signer*"]
